+++ dirigente/configurazioneUpdate.php
@@ -19,6 +19,7 @@ if(isset($_POST)) {
 	$__config->setOre_fatte_aperto($_POST['ore_fatte_aperto']);
 	$__config->setVoti_recupero_settembre_aperto($_POST['voti_recupero_settembre_aperto']);
 	$__config->setVoti_recupero_novembre_aperto($_POST['voti_recupero_novembre_aperto']);
+	$__config->setEmail_carenze_aperto($_POST['email_carenze_aperto']);
 
 	$__config->save();
 }